+++ statistics-impl/src/test/java/com/dwarfeng/statistics/impl/service/HistoryTaskEventMaintainServiceImplTest.java
@@ -44,7 +44,7 @@ public class HistoryTaskEventMaintainServiceImplTest {
 
     @Before
     public void setUp() {
-        statisticsSetting = new StatisticsSetting(new LongIdKey(STATISTICS_SETTING_ID), true, "name", "remark");
+        statisticsSetting = new StatisticsSetting(new LongIdKey(STATISTICS_SETTING_ID), true, "name","description", "remark");
         historyTask = new HistoryTask(
                 new LongIdKey(HISTORY_TASK_ID), new LongIdKey(STATISTICS_SETTING_ID), 12450, new Date(), new Date(),
                 new Date(), 12450L, new Date(), new Date(), 12450, "frontMessage", "remark"